feat(home): Enable multiline input in TextField
The `TextField` in `HomeScreen` is updated to support multiline text, allowing for up to 5 lines of input. The `singleLine` property is set to false, and a `maxLines` constraint is added.

Additionally, hardcoded dimension values for the `TextField`'s padding and height are replaced with `InputHorizontal` and `InputHeight` from the `Dimensions` object. These new dimension values are defined in `Dimensions.kt`. Several unused imports have also been removed from `HomeScreen.kt`.